requirements.txt
crewai==0.80.0
crewai-tools==0.17.0
langchain-groq==0.2.4
langchain==0.3.14
groq==0.13.0
qdrant-client==1.12.1
fastapi==0.115.6
uvicorn==0.32.1
arxiv==2.1.3
pypdf==5.1.0
sentence-transformers==3.3.1
python-dotenv==1.0.1
streamlit==1.41.1
requests==2.32.3
reportlab==4.2.5
pydantic==2.10.3
httpx==0.28.1
google-search-results==2.4.2
